+++ tests/test_http.py
@@ -8,6 +8,7 @@ import requests
 import requests_mock
 
 from yawast.scanner.plugins.http import http_basic, response_scanner
+from yawast.scanner.plugins.http.applications import wordpress
 from yawast.scanner.plugins.http.response_scanner import _check_cache_headers
 from yawast.scanner.plugins.http.servers import rails, python, nginx
 from yawast.shared import network
@@ -501,3 +502,16 @@ class TestHttpBasic(TestCase):
         res = nginx.check_banner("nginx/1.0.0", "head_data", "http://example.com")
 
         self.assertTrue(any("Nginx Outdated" in r.message for r in res))
+
+    def test_wp_path_disc(self):
+        url = "http://example.com/"
+
+        with requests_mock.Mocker() as m:
+            m.get(
+                requests_mock.ANY,
+                text="<b>Fatal error</b>:  x y() in <b>/home/akismet.php</b> on line <b>32</b><br />",
+            )
+
+            res = wordpress.check_path_disclosure(url)
+
+        self.assertTrue(any("WordPress File Path Disclosure" in r.message for r in res))